Solution to Check if a Number is Square of Another

+++ Check if a Number is Square of Another@@6788dd63490637cfc44e1ba4/code.c
@@ -2,8 +2,8 @@
 #include<math.h>
 int main()
 {
-    int a,b,c;
-    scanf("%d%d",&a,&b);
+     double int a,b,c;
+    scanf("%ld%ld",&a,&b);
     c=pow(b,2);
     if(a==c)
     printf("Yes");
